Change the way image rendering is handled
Previously, images are created and deleted on demand, but it turns out
that breaks the animation on first rendering of the images. Hopefully,
this will fix it as the images will now be created and pushed into an
array, waiting to be rendered or removed on the page. It will not
create new instances of the images, as it did previously, and will
instead reuse what's already created on the array.